Amélioration du système de Particules et créations d'effets diverses

+++ Code/ska_core/Core/GameCore.h
@@ -85,7 +85,7 @@ namespace ska {
                 	accumulator += ellapsedTime;
 
 					while (accumulator >= ti) {
-						eventUpdate(ti);
+						eventUpdate(static_cast<unsigned int>(ti));
 						accumulator -= ti;
                     }
